Backend/requirements.txt
flask~=3.1.1
gunicorn
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
psycopg2-binary
flasgger~=0.9.7.1
pyyaml~=6.0.2
werkzeug~=3.1.3
flask-cors